Changes due to rename to kebabcase
